auth_service/src/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
JWT_SECRET = os.getenv("JWT_SECRET")
JWT_ALGORITHM = os.getenv("JWT_ALGORITHM", "HS256")
JWT_EXPIRY_MINUTES = int(os.getenv("JWT_EXPIRY_MINUTES", "60"))
SERVICE_NAME = os.getenv("SERVICE_NAME", "auth_service")
LOG_DIR = os.getenv("LOG_DIR", "/var/log/auth_service")
